+++ runtime/calamari/src/xcm_config.rs
@@ -32,8 +32,8 @@ use manta_primitives::{
     assets::AssetIdLocationConvert,
     types::{AccountId, Balance, CalamariAssetId},
     xcm::{
-        AccountIdToMultiLocation, FirstAssetTrader, IsNativeConcrete, MultiAssetAdapter,
-        MultiNativeAsset,
+        AccountIdToMultiLocation, AllowTopLevelPaidExecutionDescendOriginFirst, FirstAssetTrader,
+        IsNativeConcrete, MultiAssetAdapter, MultiNativeAsset,
     },
 };
 use orml_traits::location::AbsoluteReserveProvider;
@@ -44,7 +44,7 @@ use sp_runtime::traits::Convert;
 use sp_std::prelude::*;
 use xcm::latest::prelude::*;
 use xcm_builder::{
-    AccountId32Aliases, AllowKnownQueryResponses, AllowSubscriptionsFrom,
+    Account32Hash, AccountId32Aliases, AllowKnownQueryResponses, AllowSubscriptionsFrom,
     AllowTopLevelPaidExecutionFrom, AllowUnpaidExecutionFrom, ConvertedConcreteAssetId,
     EnsureXcmOrigin, FixedRateOfFungible, LocationInverter, ParentAsSuperuser, ParentIsPreset,
     RelayChainAsNative, SiblingParachainAsNative, SiblingParachainConvertsVia,
@@ -93,6 +93,8 @@ pub type LocationToAccountId = (
     SiblingParachainConvertsVia<Sibling, AccountId>,
     // Straight up local `AccountId32` origins just alias directly to `AccountId`.
     AccountId32Aliases<RelayNetwork, AccountId>,
+    // Converts multilocation into a 32 byte hash for local `AccountId`s
+    Account32Hash<RelayNetwork, AccountId>,
 );
 
 /// This is the type to convert an (incoming) XCM origin into a local `Origin` instance,
@@ -164,6 +166,9 @@ match_types! {
 pub type Barrier = (
     // Allows local origin messages which call weight_credit >= weight_limit.
     TakeWeightCredit,
+    // Allows execution of Transact XCM instruction from configurable set of origins
+    // as long as the message is in the format DescendOrigin + WithdrawAsset + BuyExecution
+    AllowTopLevelPaidExecutionDescendOriginFirst<Everything>,
     // Allows non-local origin messages, for example from from the xcmp queue,
     // which have the ability to deposit assets and pay for their own execution.
     AllowTopLevelPaidExecutionFrom<Everything>,