+++ runtime/manta/src/xcm_config.rs
@@ -53,6 +53,8 @@ use xcm_builder::{
 };
 use xcm_executor::{traits::JustTry, Config, XcmExecutor};
 
+use manta_primitives::xcm::AllowTopLevelPaidExecutionDescendOriginFirst;
+
 parameter_types! {
     pub const ReservedDmpWeight: Weight = MAXIMUM_BLOCK_WEIGHT.saturating_div(4);
     pub const ReservedXcmpWeight: Weight = MAXIMUM_BLOCK_WEIGHT.saturating_div(4);
@@ -87,9 +89,11 @@ pub type LocationToAccountId = (
     SiblingParachainConvertsVia<Sibling, AccountId>,
     // Straight up local `AccountId32` origins just alias directly to `AccountId`.
     AccountId32Aliases<RelayNetwork, AccountId>,
+    // Converts multilocation into a 32 byte hash for local `AccountId`s
+    xcm_builder::Account32Hash<RelayNetwork, AccountId>,
 );
 
-/// This is the type to convert an (incoming) XCM origin into a local `RuntimeOrigin` instance,
+/// This is the type to convert an (incoming) XCM origin into a local `Origin` instance,
 /// ready for dispatching a transaction with Xcm's `Transact`.
 /// It uses some Rust magic macro to do the pattern matching sequentially.
 /// There is an `OriginKind` which can biases the kind of local `RuntimeOrigin` it will become.
@@ -153,6 +157,9 @@ match_types! {
 pub type Barrier = (
     // Allows local origin messages which call weight_credit >= weight_limit.
     TakeWeightCredit,
+    // Allows execution of Transact XCM instruction from configurable set of origins
+    // as long as the message is in the format DescendOrigin + WithdrawAsset + BuyExecution
+    AllowTopLevelPaidExecutionDescendOriginFirst<Everything>,
     // Allows non-local origin messages, for example from from the xcmp queue,
     // which have the ability to deposit assets and pay for their own execution.
     AllowTopLevelPaidExecutionFrom<Everything>,