+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v4.3.0
 ### Added
-- [\#1179](https://github.com/Manta-Network/Manta/pull/1179) Name Service Pallet [CA]
+- [\#1179](https://github.com/Manta-Network/Manta/pull/1179) Copy Name Service Pallet [CA]
 - [\#1141](https://github.com/Manta-Network/Manta/pull/1141) manta-farming [MACA]
 - [\#1036](https://github.com/Manta-Network/Manta/pull/1036) Pallet Staking Lottery [MACA]
 
@@ -11,6 +11,8 @@
 
 ### Fixed
 - [\#1175](https://github.com/Manta-Network/Manta/pull/1175) Fix congestion test CI [MACA]
+- [\#1193](https://github.com/Manta-Network/Manta/pull/1193) Revert to v0942 polkadot binary [MACA]
+- [\#1196](https://github.com/Manta-Network/Manta/pull/1196) Fix dex ed [MA]
 
 ### Removed
 - [\#1167](https://github.com/Manta-Network/Manta/pull/1167) Retire dolphin runtime [DO]